+++ examples/react-app/src/Player.jsx
@@ -1,4 +1,5 @@
 import "./Player.css";
+import "ezuikit-flv/style.css";
 import { useCallback, useEffect, useRef } from "react";
 import EzuikitFlv from "ezuikit-flv";
 
@@ -24,7 +25,7 @@ function Player() {
         }
 
         value = parseInt((value * 100 + "").split(".")[0]) / 100; // 不使用 toFixed 是为了避免四舍五入问题
-        playerRef.current.setVolume(value);
+        playerRef.current.volume = value;
       } else {
         console.log("player 未初始化");
       }
@@ -38,10 +39,13 @@ function Player() {
         container: containerRef.current,
         debug: true,
         url,
-        useMSE: true,
-        decoder: "decoder.js" // 软解解码资源 （wasm 要和js 在同一个文件夹中）
+        useMSE: true, // 硬解
+        autoPlay: true, // 默认自动播放
+        // decoder: "decoder.js", // 软解解码资源 （wasm 要和js 在同一个文件夹中）
       });
-      playerRef.current.play();
+      // 调试api 使用
+      window.player = playerRef.current;
+      // playerRef.current.play();
     }
   };
 
@@ -65,6 +69,12 @@ function Player() {
     }
   }, []);
 
+  const handleScreenshot = useCallback(() => {
+    if (playerRef.current) {
+      playerRef.current.screenshot();
+    }
+  }, []);
+
   const handleDestroy = useCallback(() => {
     if (playerRef.current) {
       playerRef.current.destroy();
@@ -74,25 +84,26 @@ function Player() {
 
   const handleOpenSound = useCallback(() => {
     if (playerRef.current) {
-      playerRef.current.openSound();
+      playerRef.current.muted = false
+      playerRef.current.volume = 0.8
     }
   }, []);
 
   const handleCloseSound = useCallback(() => {
     if (playerRef.current) {
-      playerRef.current.closeSound();
+      playerRef.current.muted = true
     }
   }, []);
 
   const handleFullScreen = useCallback(() => {
     if (playerRef.current) {
-      playerRef.current.fullScreen();
+      playerRef.current.fullscreen();
     }
   }, []);
 
   const handleCancelFullScreen = useCallback(() => {
     if (playerRef.current) {
-      playerRef.current.cancelFullScreen();
+      playerRef.current.exitFullscreen();
     }
   }, []);
 
@@ -118,6 +129,7 @@ function Player() {
           <button onClick={handleInIt}>init</button>
           <button onClick={handlePlay}>播放</button>
           <button onClick={handlePause}>暂停</button>
+          <button onClick={handleScreenshot}>截图</button>
           <button onClick={handleOpenSound}>打开声音</button>
           <button onClick={handleCloseSound}>关闭声音</button>
           <button onClick={handleFullScreen}>开启全屏</button>