+++ examples/vue-app/src/App.vue
@@ -3,7 +3,5 @@ import Player from './components/Player.vue'
 </script>
 
 <template>
-  <main>
     <Player />
-  </main>
 </template>