+++ examples/vue-app/src/components/Player.vue
@@ -1,4 +1,5 @@
 <script setup>
+import 'ezuikit-flv/style.css'
 import { onMounted } from 'vue'
 import EzuikitFlv from 'ezuikit-flv'
 
@@ -8,24 +9,22 @@ onMounted(() => {
   player = new EzuikitFlv({
     container: document.getElementById('container'),
     debug: true,
-    url: 'https://test12flow.ys7.com:443/v3/openlive/C91672728_1_2.flv?expire=1724231462&id=615956453163966465&t=3edca10be26e834f298923d4114bf9467d9778a7e0d1fa02342846031545b909&ev=100',
+    url: 'https://rtmp05open.ys7.com:9188/v3/openlive/BC7799091_1_1.flv?expire=1758815004&id=891462424098512896&t=09d091ee55513927407935cd1ce038e44e7e05029ebd5d57f8ec0e1be959aadf&ev=100&supportH265=1',
     useMSE: true,
-    decoder: 'decoder.js' // 软解解码资源 （wasm 要和js 在同一个文件夹中）
+    autoPlay: true,
+    // decoder: 'decoder.js' // 软解解码资源 （wasm 要和js 在同一个文件夹中）
   })
-  player.play()
+  // player.play()
 })
 </script>
 
 <template>
-  <div>
     <div id="container"></div>
-  </div>
 </template>
 
 <style>
 #container {
-  width: 600px;
   height: 400px;
-  background: #000;
+  width: 100%;
 }
 </style>